+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 2020-01-19 - Release 1.0.0-rc1
+* Greatly improved `colspan` & `rowspan` support.
+* When using `includeRowId`, `rowid` is now set to `null` if the row does not have an ID set (previously gave an empty string).
+* More thorough tests.
+* More modular codebase.
+
 ## 2020-01-17 - Release 0.13.1
 * `ignoreEmptyRows` feature should consider `ignoreColumns` when deciding what is considered an empty row.